test: reorganize ITs and remove unused fields
